chore(admin-cli): migrate `sku` command group to `Dispatch` + `Run`
We recently introduced a `Dispatch` trait that is used by both `main.rs` for dispatching to command groups (which can also dispatch to nested command groups), and now we have a `Run` trait for dispatch handlers to run the actual command, which was introduced in [this PR](https://github.com/NVIDIA/bare-metal-manager-core/pull/375), which explains what is being done here.

That initial PR only pulled in the `credential` command group to keep the PR easier to understand. I'm now going to start pulling in other command group refactors to fit the pattern. I'll probably make the PRs bigger and bigger as I do more and more, but it's nice to be able to look at the single-group refactors for what they are.
