+++ crates/admin-cli/src/sku/assign/cmd.rs
@@ -1,0 +1,39 @@
+/*
+ * SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+ * SPDX-License-Identifier: Apache-2.0
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+use ::rpc::admin_cli::CarbideCliResult;
+use carbide_uuid::machine::MachineId;
+use rpc::forge::SkuMachinePair;
+
+use crate::rpc::ApiClient;
+
+pub async fn assign(
+    sku_id: String,
+    machine_id: MachineId,
+    force: bool,
+    api_client: &ApiClient,
+) -> CarbideCliResult<()> {
+    api_client
+        .0
+        .assign_sku_to_machine(SkuMachinePair {
+            sku_id,
+            machine_id: Some(machine_id),
+            force,
+        })
+        .await?;
+    Ok(())
+}
