+++ crates/admin-cli/src/sku/delete/cmd.rs
@@ -1,0 +1,29 @@
+/*
+ * SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+ * SPDX-License-Identifier: Apache-2.0
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+use ::rpc::admin_cli::CarbideCliResult;
+use rpc::forge::SkuIdList;
+
+use crate::rpc::ApiClient;
+
+pub async fn delete(sku_id: String, api_client: &ApiClient) -> CarbideCliResult<()> {
+    api_client
+        .0
+        .delete_sku(SkuIdList { ids: vec![sku_id] })
+        .await?;
+    Ok(())
+}
