+++ crates/admin-cli/src/sku/replace/args.rs
@@ -1,0 +1,29 @@
+/*
+ * SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+ * SPDX-License-Identifier: Apache-2.0
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+use clap::Parser;
+
+use super::super::common::CreateSkuOptions;
+
+// Args wraps the shared CreateSkuOptions as a subcommand
+// specific newtype to allow sharing of CreateSkuOptions, and still
+// providing a subcommand-specific Run trait implementation.
+#[derive(Parser, Debug)]
+pub struct Args {
+    #[clap(flatten)]
+    pub inner: CreateSkuOptions,
+}
